@@ -3,10 +3,10 @@ description: Social app for tasrufdars
 
 publish_to: 'none' 
 
-version: 1.6.0+16
+version: 1.7.0+17
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,8 +14,8 @@ dependencies:
 
   cupertino_icons: ^1.0.5
   form_field_validator: ^1.1.0
-  flutter_svg: ^1.0.3
-  pinput: ^2.2.21
+  flutter_svg: ^2.0.9
+  pinput: ^3.0.1
   about: ^2.1.1
   flutter_vector_icons: ^2.0.0
   share: ^2.0.4
@@ -25,17 +25,18 @@ dependencies:
   awesome_dialog: ^3.0.2
   provider: ^6.0.5
   shared_preferences: ^2.0.15
-  upgrader: ^4.8.1
+  upgrader: ^8.3.0
   swipe_to: ^1.0.2
-  flutter_linkify: ^5.0.2
+  flutter_linkify: ^6.0.0
   animated_background: ^2.0.0
+  url_launcher: ^6.2.1
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_native_splash: ^1.1.8+4
-  flutter_lints: ^2.0.1
+  flutter_native_splash: ^2.3.6
+  flutter_lints: ^3.0.1
 
 flutter_native_splash:
   image: assets/images/splash.png
